Removing Ting.py, updating __init__.py for setuptools - adding contributor / author
